Refactoring and battle plan for moving our GraphicManager into 2 separate classes or more perhaps.
We need to be able to share our factories with multiple HWND.

Also thinking about how I can have it know about animated GIFs, and have a timer to render multiple frames at a speed of 30 fps or so.
I do not know if we have a way to gather information about the FPS available in the GIFs?